feat: main으로 주소 이동
