+++ pages/main.tsx
@@ -1,0 +1,38 @@
+import Header from '@/components/Header';
+import DataTabs from '@/components/Main/cakeTabs';
+
+export default function Home() {
+  const tabData = [
+    {
+      label: '전체',
+      content: '전체 케이크',
+    },
+    {
+      label: '생크림',
+      content: '생크림 케이크',
+    },
+    {
+      label: '쵸코',
+      content: '초콜릿 케이크',
+    },
+    {
+      label: '전체',
+      content: '전체 케이크',
+    },
+    {
+      label: '생크림',
+      content: '생크림 케이크',
+    },
+    {
+      label: '쵸코',
+      content: '초콜릿 케이크',
+    },
+  ];
+
+  return (
+    <>
+      <Header />
+      <DataTabs data={tabData} />
+    </>
+  );
+}
